+++ src/datasources/asset/types/ListAssets.types.ts
@@ -1,8 +1,9 @@
 import { DataQuery } from '@grafana/schema'
+import { QueryBuilderOperations } from '../../../core/query-builder.constants';
+import { QBField } from '../../asset-calibration/types';
 
 export interface ListAssetsQuery extends DataQuery {
-    workspace: string,
-    minionIds: string[]
+    filter: string;
 }
 
 export enum AssetFilterProperties {
@@ -33,8 +34,34 @@ export enum AssetFilterProperties {
     ExternalCalibrationComments = 'ExternalCalibration.Comments',
     ExternalCalibrationIsLimited = 'ExternalCalibration.IsLimited',
     ExternalCalibrationOperatorDisplayName = 'ExternalCalibration.Operator.DisplayName',
-    IsSystemController = 'IsSystemController'
+    IsSystemController = 'IsSystemController',
+    Workspace = 'Workspace'
 }
+
+export const ListAssetsFields: Record<string, QBField> = {
+    LOCATION: {
+        label: 'Location',
+        dataField: AssetFilterProperties.LocationMinionId,
+        filterOperations: [
+            QueryBuilderOperations.EQUALS.name,
+            QueryBuilderOperations.DOES_NOT_EQUAL.name
+        ],
+        lookup: {
+            dataSource: []
+        }
+    },
+    WORKSPACE: {
+        label: 'Workspace',
+        dataField: AssetFilterProperties.Workspace,
+        filterOperations: [
+            QueryBuilderOperations.EQUALS.name,
+            QueryBuilderOperations.DOES_NOT_EQUAL.name
+        ],
+        lookup: {
+            dataSource: []
+        }
+    }
+};
 
 export enum EntityType {
     Asset = "Asset",